annotation_server/requirements.txt
djangorestframework==3.11.0
django==3.2.18
asgiref==3.3.2
pip==23.0
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.6
wheel==0.37.1
django-taggit==1.2.0
psycopg2==2.9.5

# TODO: Add development packages here
flake8
coverage
ipdb
pre-commit
django-debug-toolbar
django-extensions
django-coverage-plugin
